move log to top level dir
