+++ kv/raftstore/runner/raftlog_gc.go
@@ -2,7 +2,7 @@ package runner
 
 import (
 	"github.com/Connor1996/badger"
-	"github.com/pingcap-incubator/tinykv/kv/util/log"
+	"github.com/pingcap-incubator/tinykv/log"
 	"github.com/pingcap-incubator/tinykv/kv/raftstore/meta"
 	"github.com/pingcap-incubator/tinykv/kv/util/engine_util"
 	"github.com/pingcap-incubator/tinykv/kv/worker"
