+++ raft/raft.go
@@ -23,7 +23,7 @@ import (
 	"sync"
 	"time"
 
-	"github.com/pingcap-incubator/tinykv/kv/util/log"
+	"github.com/pingcap-incubator/tinykv/log"
 	pb "github.com/pingcap-incubator/tinykv/proto/pkg/eraftpb"
 )
